Initial code for RBAC
Involves holding a number of admin Roles and can be expanded for judge roles also.

admins, so far restrict the starting of the competition.
Can be further restrictions so only an admin can finish or only a judge can pay for example.
Added 3 new tests; this and existing tests means Roles are tested
Formatted test file
Test in Remix